requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
cryptography==46.0.3
cssselect2==0.8.0
Django==4.2.7
django-extensions==3.2.3
django-ratelimit==4.1.0
et_xmlfile==2.0.0
greenlet==3.3.0
gunicorn==21.2.0
html5lib==1.1
idna==3.11
joblib==1.5.3
lxml==6.0.2
numpy==2.4.0
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.3.3
pillow==10.3.0
playwright==1.55.0
psycopg2-binary==2.9.9
pycparser==2.23
pyee==13.0.0
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pypdf==6.4.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
reportlab==4.0.9
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sqlparse==0.5.4
svglib==1.5.1
threadpoolctl==3.6.0
tinycss2==1.5.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.6.0
webencodings==0.5.1
whitenoise==6.11.0
xhtml2pdf==0.2.15
google-generativeai